Move all command cli logic to a new module
