+++ cli/src/command/add_command.rs
