+++ cli/src/command/mod.rs
@@ -1,0 +1,7 @@
+pub mod add_command;
+pub mod delete_command;
+pub mod list_commands;
+pub mod search_command;
+pub mod search_utils;
+pub mod update_command;
+
