+++ cli/src/command/search_utils.rs
