+++ cli/src/param/mod.rs
@@ -1,6 +1,6 @@
 ///! Handle all commands relating to parameters
 use crate::args::ParamCommands;
-use crate::search_utils::{
+use crate::command::search_utils::{
     display_search_args_wizard, get_searched_commands, search_args_wizard,
     GetSelectedItemFromUserError,
 };
